Require a poller name in `with-zmq-poller`
Account for the possibility of multiple poller objects. The current
implementation somewhat assumes that there exists only one poller object when
running the body of `with-zmq-poller`. Requiring a poller name allows for
multiple nested calls of `with-zmq-poller` each with a different poller object.

This removes `current-zmq-poller` since you can now access the poller object by
the name supplied to `with-zmq-poller`