Add global reference stack for releasing global references
This is in preparation for removing the `refcount` slot of an `ezmq_obj_t`. It
is needed since there needs to be a way to release the global references
created in `ezmq_obj_set_val` when `ezmq_free_obj` when ZMQ objects are garbage
collected. At garbage collection time, there is no valid Emacs environment
pointer available, thus we need to defer the release of the global references
until one is available.

To do this, the global references flagged for release are cleaned up at the end
of every `ezmq_dispatch` call. In addition, a hook function is added to
`post-gc-hook` to do cleanup. This hook is needed since we also need to handle
the case where a global reference is flagged for release during garbage
collection, but no more calls to zmq functions are made afterwards. In this
case, the cleanup in `ezmq_dispatch` does not occur and the references flagged
for release are never released.