Fix polling tests
The polling tests don't fail intermittently anymore. The issue seemed to be
with setting the message subscribe filter after waiting for the sockets to
connect. We were not allowing the `zmq-PUB` socket to receive the subscription
message generated by setting the `zmq-SUBSCRIBE` option for `zmq-SUB` socket.
Waiting after sending the subscription message allows the connection between
the two sockets to happen consistently.