zmq-load: Be more robust when loading the module
Emacs >= 28 supports two types of module file suffixes on macOS so we
let Emacs do the checking if there is an appropriate module file
instead of constructing a path based on `module-file-suffix`.  This
amounts to using `load` instead of `load-file`, the latter doesn't
check if there is a matching file with a suffix in `load-suffixes`.